integrated Post & Countdown into NewItems
