+++ src/components/home/NewItems.jsx
@@ -1,122 +1,7 @@
-import React, { useEffect, useState } from "react";
-import { Link } from "react-router-dom";
-import axios from "axios";
-import Skeleton from "../UI/Skeleton";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-import Slider from "react-slick";
+import React from "react";
+import Post from "../UI/Post";
 
 const NewItems = () => {
-  const [newItems, setNewItems] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [settings, setSettings] = useState([]);
-
-  useEffect(() => {
-    async function fetchNewItems() {
-      const { data } = await axios.get(
-        "https://us-central1-nft-cloud-functions.cloudfunctions.net/newItems"
-      );
-      setNewItems(data);
-      setLoading(false);
-    }
-    fetchNewItems();
-  }, []);
-
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setNewItems((prevItems) =>
-        prevItems.map((item) => {
-          if (item.expiryDate) {
-            const timeLeft = item.expiryDate - Date.now();
-            if (timeLeft > 0) {
-              return {
-                ...item,
-                timeLeft: {
-                  hours: Math.floor((timeLeft / (1000 * 60 * 60)) % 24),
-                  minutes: Math.floor((timeLeft / (1000 * 60)) % 60),
-                  seconds: Math.floor((timeLeft / 1000) % 60),
-                },
-              };
-            } else {
-              return {
-                ...item,
-                timeLeft: "Expired",
-              };
-            }
-          } else {
-            return {
-              ...item,
-              timeLeft: null,
-            };
-          }
-        })
-      );
-    }, 1000);
-
-    return () => clearInterval(interval);
-  }, [newItems]);
-
-  useEffect(() => {
-    function SimpleSlider() {
-      const options = {
-        infinite: true,
-        speed: 500,
-        slidesToShow: 4,
-        slidesToScroll: 1,
-        nextArrow: <NextArrow />,
-        prevArrow: <PrevArrow />,
-        responsive: [
-          {
-          breakpoint: 1200,
-          settings: {
-            slidesToShow: 3,
-            slidesToScroll: 1,
-            infinite: true,
-          }
-        },
-        {
-          breakpoint: 992,
-          settings: {
-            slidesToShow: 2,
-            slidesToScroll: 1,
-            infinite: true,
-          }
-        },
-        {
-          breakpoint: 768,
-          settings: {
-            slidesToShow: 1,
-            slidesToScroll: 1,
-            infinite: true,
-          }
-        }]
-      };
-      setSettings(options);
-    }
-    SimpleSlider();
-  }, []);
-
-  function NextArrow(props) {
-    const { className, style, onClick } = props;
-    return (
-      <div
-        className={className}
-        style={{ ...style, display: "flex", position: "absolute" }}
-        onClick={onClick}
-      />
-    );
-  }
-
-  function PrevArrow(props) {
-    const { className, style, onClick } = props;
-    return (
-      <div
-        className={className}
-        style={{ ...style, display: "flex", position: "absolute" }}
-        onClick={onClick}
-      />
-    );
-  }
 
   return (
     <section id="section-items" className="no-bottom">
@@ -128,93 +13,7 @@ const NewItems = () => {
               <div className="small-border bg-color-2"></div>
             </div>
           </div>
-          <Slider {...settings}>
-              {loading
-                ? new Array(4).fill(0).map((_, index) => (
-                    <div className="nft__item" key={index}>
-                      <div className="author_list_pp">
-                        <Skeleton
-                          width="50px"
-                          height="50px"
-                          border-radius="50%"
-                        />
-                      </div>
-                      <div className="nft__item_wrap">
-                        <div className="nft__item_extra"></div>
-                        <Skeleton width="100%" height="335px" />
-                      </div>
-                      <div className="nft__item_info">
-                        <Skeleton width="180px" height="30px" />
-                        <div className="nft__item_price">
-                          <Skeleton width="100px" height="20px" />
-                        </div>
-                        <div className="nft__item_like">
-                          <Skeleton width="30px" height="15px" />
-                        </div>
-                      </div>
-                    </div>
-                  ))
-                : newItems.map((item, index) => (
-                    <div className="nft__item">
-                      <div className="author_list_pp">
-                        <Link
-                          to="/author"
-                          data-bs-toggle="tooltip"
-                          data-bs-placement="top"
-                          title="Creator: Monica Lucas"
-                        >
-                          <img className="lazy" src={item.authorImage} alt="" />
-                          <i className="fa fa-check"></i>
-                        </Link>
-                      </div>
-
-                      {item.expiryDate && item.timeLeft ? (
-                        <div className="de_countdown">
-                          {item.timeLeft.hours}h {item.timeLeft.minutes}m{" "}
-                          {item.timeLeft.seconds}s
-                        </div>
-                      ) : null}
-
-                      <div className="nft__item_wrap">
-                        <div className="nft__item_extra">
-                          <div className="nft__item_buttons">
-                            <button>Buy Now</button>
-                            <div className="nft__item_share">
-                              <h4>Share</h4>
-                              <a href="" target="_blank" rel="noreferrer">
-                                <i className="fa fa-facebook fa-lg"></i>
-                              </a>
-                              <a href="" target="_blank" rel="noreferrer">
-                                <i className="fa fa-twitter fa-lg"></i>
-                              </a>
-                              <a href="">
-                                <i className="fa fa-envelope fa-lg"></i>
-                              </a>
-                            </div>
-                          </div>
-                        </div>
-
-                        <Link to="/item-details">
-                          <img
-                            src={item.nftImage}
-                            className="lazy nft__item_preview"
-                            alt=""
-                          />
-                        </Link>
-                      </div>
-                      <div className="nft__item_info">
-                        <Link to="/item-details">
-                          <h4>{item.title}</h4>
-                        </Link>
-                        <div className="nft__item_price">{item.price} ETH</div>
-                        <div className="nft__item_like">
-                          <i className="fa fa-heart"></i>
-                          <span>{item.likes}</span>
-                        </div>
-                      </div>
-                    </div>
-                  ))}
-          </Slider>
+          <Post />
         </div>
       </div>
     </section>
